pyproject.toml
[project]
name = "semantic-search-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.0",
    "langchain-community>=0.0.20",
    "langchain-core>=1.1.0",
    "langchain-google-genai>=1.0.0",
    "langchain-pinecone>=0.2.13",
    "pinecone>=7.3.0",
    "plotly>=6.5.0",
    "pypdf>=3.17.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
